microservices/activity-service/requirements.txt
aws-msk-iam-sasl-signer-python==1.0.2
blinker==1.9.0
boto3==1.34.0
botocore==1.34.162
click==8.3.1
Flask==3.0.0
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kafka-python==2.3.0
MarkupSafe==3.0.3
packaging==25.0
prometheus-flask-exporter==0.23.0
prometheus_client==0.23.1
python-dateutil==2.9.0.post0
s3transfer==0.9.0
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
